requirements.txt
# Core ML and NLP
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0

# Data processing
pandas==2.0.3
pyyaml==6.0.1

# CLI and utilities
click==8.1.7
python-dotenv==1.0.0

# Monitoring and system utilities
psutil==5.9.5

# Development and testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Optional: GPU support (uncomment if needed)
# faiss-gpu==1.7.4
# torch==2.0.1 